ci: docker build files
Two docker containers.
First contains all the necessary dependencies for our build environment.
Second build the project and runs tests

The build container seem huge (over 100GB), probably because of the size of LLVM
debug build. We might not be able to easily use in a CI. We will see :) ...